split server actions into fetch and mutation
